[feat] probID 의 타입 Long --> Integer 로 수정

+++ eTest/src/main/java/com/tmax/eTest/Contents/model/DiagnosisProblem.java
@@ -17,7 +17,7 @@ public class DiagnosisProblem {
 	
 	@Id
 	@Column(name="PROB_ID")
-	private long probID;
+	private Integer probID;
 	
 	@Column(name="SUBJECT")
 	private int subject;
